# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Incognito Intel API - Anonymous student Q&A platform with teacher analytics
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Question {
  id: number;
  content: string;
  /** @nullable */
  topicId: number | null;
  /** @nullable */
  topicName: string | null;
  /** @nullable */
  topicCategory: string | null;
  addressed: boolean;
  /** @nullable */
  answer: string | null;
  /** @nullable */
  answeredAt: string | null;
  createdAt: string;
}

export interface QuestionUpdate {
  addressed?: boolean;
  /** @nullable */
  answer?: string | null;
}

export interface QuestionInput {
  /** @minLength 5 */
  content: string;
  /** @nullable */
  topicId?: number | null;
}

export interface Topic {
  id: number;
  name: string;
  category: string;
  questionCount: number;
}

export interface TopicInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  category: string;
}

export interface HeatmapEntry {
  topicId: number;
  topicName: string;
  category: string;
  count: number;
  percentage: number;
}

export interface AnalyticsSummary {
  totalQuestions: number;
  totalTopics: number;
  /** @nullable */
  topTopic: string | null;
  /** @nullable */
  topTopicCount: number | null;
  /** @nullable */
  topTopicPercentage: number | null;
  questionsToday: number;
  questionsThisWeek: number;
}

export type ListQuestionsParams = {
  topic?: string;
  limit?: number;
  offset?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};