# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Incognito Intel API - Anonymous student Q&A platform with teacher analytics
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  content: string;
  /** @nullable */
  topicId: number | null;
  /** @nullable */
  topicName: string | null;
  /** @nullable */
  topicCategory: string | null;
  addressed: boolean;
  /** @nullable */
  answer: string | null;
  /** @nullable */
  answeredAt: string | null;
  createdAt: string;
}