# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Incognito Intel API - Anonymous student Q&A platform with teacher analytics
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  addressed?: boolean;
  /** @nullable */
  answer?: string | null;
}